support for bigger packetsize (>4095)

+++ src/tp/isotp-c/isotp.h
@@ -24,12 +24,12 @@ typedef struct IsoTpLink {
     uint32_t                    send_arbitration_id; /* used to reply consecutive frame */
     /* message buffer */
     uint8_t*                    send_buffer;
-    uint16_t                    send_buf_size;
-    uint16_t                    send_size;
-    uint16_t                    send_offset;
+    uint32_t                    send_buf_size;
+    uint32_t                    send_size;
+    uint32_t                    send_offset;
     /* multi-frame flags */
     uint8_t                     send_sn;
-    uint16_t                    send_bs_remain; /* Remaining block size */
+    uint32_t                    send_bs_remain; /* Remaining block size */
     uint32_t                    send_st_min_us; /* Separation Time between consecutive frames */
     uint8_t                     send_wtf_count; /* Maximum number of FC.Wait frame transmissions  */
     uint32_t                    send_timer_st;  /* Last time send consecutive frame */    
@@ -42,9 +42,9 @@ typedef struct IsoTpLink {
     uint32_t                    receive_arbitration_id;
     /* message buffer */
     uint8_t*                    receive_buffer;
-    uint16_t                    receive_buf_size;
-    uint16_t                    receive_size;
-    uint16_t                    receive_offset;
+    uint32_t                    receive_buf_size;
+    uint32_t                    receive_size;
+    uint32_t                    receive_offset;
     /* multi-frame control */
     uint8_t                     receive_sn;
     uint8_t                     receive_bs_count; /* Maximum number of FC.Wait frame transmissions  */
@@ -70,8 +70,8 @@ typedef struct IsoTpLink {
  * @param recvbufsize The size of the buffer area.
  */
 void isotp_init_link(IsoTpLink *link, uint32_t sendid, 
-                     uint8_t *sendbuf, uint16_t sendbufsize,
-                     uint8_t *recvbuf, uint16_t recvbufsize);
+                     uint8_t *sendbuf, uint32_t sendbufsize,
+                     uint8_t *recvbuf, uint32_t recvbufsize);
 
 /**
  * @brief Polling function; call this function periodically to handle timeouts, send consecutive frames, etc.
@@ -106,12 +106,12 @@ void isotp_on_can_message(IsoTpLink *link, const uint8_t *data, uint8_t len);
  *  - @code ISOTP_RET_OK @endcode
  *  - The return value of the user shim function isotp_user_send_can().
  */
-int isotp_send(IsoTpLink *link, const uint8_t payload[], uint16_t size);
+int isotp_send(IsoTpLink *link, const uint8_t payload[], uint32_t size);
 
 /**
  * @brief See @link isotp_send @endlink, with the exception that this function is used only for functional addressing.
  */
-int isotp_send_with_id(IsoTpLink *link, uint32_t id, const uint8_t payload[], uint16_t size);
+int isotp_send_with_id(IsoTpLink *link, uint32_t id, const uint8_t payload[], uint32_t size);
 
 /**
  * @brief Receives and parses the received data and copies the parsed data in to the internal buffer.
@@ -124,7 +124,7 @@ int isotp_send_with_id(IsoTpLink *link, uint32_t id, const uint8_t payload[], ui
  *      - @link ISOTP_RET_OK @endlink
  *      - @link ISOTP_RET_NO_DATA @endlink
  */
-int isotp_receive(IsoTpLink *link, uint8_t *payload, const uint16_t payload_size, uint16_t *out_size);
+int isotp_receive(IsoTpLink *link, uint8_t *payload, const uint32_t payload_size, uint32_t *out_size);
 
 #ifdef __cplusplus
 }
